Separated by id and leetcodeUsername endpoints into different controllers
